Add additional categories to tiling
It is now possible to add additional categories to the category
tiling in the category page. The default contains all the
subcategories of the current category, which can now be extended
with additional categories that the admin can select himself.

Also a testing suite is implemented, which adds additional coding
standard checks.